+++ indispensable-cosef/src/commonTest/kotlin/at/asitplus/signum/indispensable/cosef/CoseEqualsTest.kt
@@ -2,7 +2,6 @@ package at.asitplus.signum.indispensable.cosef
 
 import at.asitplus.signum.indispensable.CryptoSignature
 import at.asitplus.signum.indispensable.cosef.io.Base16Strict
-import at.asitplus.signum.indispensable.cosef.io.coseCompliantSerializer
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
@@ -13,24 +12,24 @@ import io.kotest.property.arbitrary.int
 import io.kotest.property.checkAll
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.encodeToByteArray
+import kotlinx.serialization.builtins.ByteArraySerializer
 
 class CoseEqualsTest : FreeSpec({
     "equals with byte array" {
         checkAll(Arb.byteArray(length = Arb.int(0, 10), content = Arb.byte())) { bytes ->
-            val bytesSigned1 = CoseSigned<ByteArray>(
+            val bytesSigned1 = CoseSigned.create(
                 protectedHeader = CoseHeader(),
                 unprotectedHeader = null,
                 payload = bytes,
                 signature = CryptoSignature.RSAorHMAC(bytes),
-                rawPayload = bytes,
+                payloadSerializer = ByteArraySerializer(),
             )
-            val bytesSigned2 = CoseSigned<ByteArray>(
+            val bytesSigned2 = CoseSigned.create(
                 protectedHeader = CoseHeader(),
                 unprotectedHeader = null,
                 payload = bytes,
                 signature = CryptoSignature.RSAorHMAC(bytes),
-                rawPayload = bytes,
+                payloadSerializer = ByteArraySerializer(),
             )
 
             bytesSigned1 shouldBe bytesSigned1
@@ -39,19 +38,19 @@ class CoseEqualsTest : FreeSpec({
             bytesSigned1.hashCode() shouldBe bytesSigned2.hashCode()
 
             val reversed = bytes.reversedArray().let { it + it + 1 + 3 + 5 }
-            val reversedSigned1 = CoseSigned<ByteArray>(
+            val reversedSigned1 = CoseSigned.create(
                 protectedHeader = CoseHeader(),
                 unprotectedHeader = null,
                 payload = reversed,
                 signature = CryptoSignature.RSAorHMAC(reversed),
-                rawPayload = reversed,
+                payloadSerializer = ByteArraySerializer(),
             )
-            val reversedSigned2 = CoseSigned<ByteArray>(
+            val reversedSigned2 = CoseSigned.create(
                 protectedHeader = CoseHeader(),
                 unprotectedHeader = null,
                 payload = reversed,
                 signature = CryptoSignature.RSAorHMAC(reversed),
-                rawPayload = reversed,
+                payloadSerializer = ByteArraySerializer(),
             )
 
             reversedSigned2 shouldBe reversedSigned2
@@ -77,19 +76,19 @@ class CoseEqualsTest : FreeSpec({
     "equals with data class" {
         checkAll(Arb.byteArray(length = Arb.int(0, 10), content = Arb.byte())) { bytes ->
             val payload = DataClass(content = bytes.encodeToString(Base16Strict))
-            val bytesSigned1 = CoseSigned(
+            val bytesSigned1 = CoseSigned.create(
                 protectedHeader = CoseHeader(),
                 unprotectedHeader = null,
                 payload = payload,
                 signature = CryptoSignature.RSAorHMAC(bytes),
-                rawPayload = coseCompliantSerializer.encodeToByteArray(payload),
+                payloadSerializer = DataClass.serializer(),
             )
-            val bytesSigned2 = CoseSigned(
+            val bytesSigned2 = CoseSigned.create(
                 protectedHeader = CoseHeader(),
                 unprotectedHeader = null,
                 payload = payload,
                 signature = CryptoSignature.RSAorHMAC(bytes),
-                rawPayload = coseCompliantSerializer.encodeToByteArray(payload),
+                payloadSerializer = DataClass.serializer(),
             )
 
             bytesSigned1 shouldBe bytesSigned1
@@ -99,19 +98,19 @@ class CoseEqualsTest : FreeSpec({
 
             val reversed =
                 DataClass(content = bytes.reversedArray().let { it + it + 1 + 3 + 5 }.encodeToString(Base16Strict))
-            val reversedSigned1 = CoseSigned(
+            val reversedSigned1 = CoseSigned.create(
                 protectedHeader = CoseHeader(),
                 unprotectedHeader = null,
                 payload = reversed,
                 signature = CryptoSignature.RSAorHMAC(bytes),
-                rawPayload = coseCompliantSerializer.encodeToByteArray(reversed),
+                payloadSerializer = DataClass.serializer(),
             )
-            val reversedSigned2 = CoseSigned(
+            val reversedSigned2 = CoseSigned.create(
                 protectedHeader = CoseHeader(),
                 unprotectedHeader = null,
                 payload = reversed,
                 signature = CryptoSignature.RSAorHMAC(bytes),
-                rawPayload = coseCompliantSerializer.encodeToByteArray(reversed),
+                payloadSerializer = DataClass.serializer(),
             )
 
             reversedSigned2 shouldBe reversedSigned2