+++ indispensable-cosef/src/commonTest/kotlin/at/asitplus/signum/indispensable/cosef/CoseSerializationTest.kt
@@ -12,51 +12,38 @@ import io.kotest.matchers.string.shouldContain
 import io.matthewnelson.encoding.base16.Base16
 import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
+import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.builtins.ByteArraySerializer
+import kotlinx.serialization.builtins.serializer
 import kotlinx.serialization.encodeToByteArray
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import kotlin.random.Random
 
+@OptIn(ExperimentalSerializationApi::class)
 class CoseSerializationTest : FreeSpec({
 
     "CoseSigned can not be constructed with ByteStringWrapper" {
         val payload = ByteStringWrapper("StringType")
         shouldThrow<IllegalArgumentException> {
-            CoseSigned<ByteStringWrapper<*>>(
+            CoseSigned.create(
                 protectedHeader = CoseHeader(algorithm = CoseAlgorithm.ES256),
                 unprotectedHeader = CoseHeader(),
                 payload = payload,
                 signature = CryptoSignature.RSAorHMAC(byteArrayOf()),
-                rawPayload = null,
+                payloadSerializer = ByteStringWrapper.serializer(String.serializer())
             )
         }
     }
 
-    "Serialization is correct for byte array" {
-        val payload = "This is the content.".encodeToByteArray()
-        val cose = CoseSigned<ByteArray>(
-            protectedHeader = CoseHeader(algorithm = CoseAlgorithm.ES256),
-            unprotectedHeader = CoseHeader(),
-            payload = payload,
-            signature = CryptoSignature.RSAorHMAC("bar".encodeToByteArray()), //RSAorHMAC because EC expects tuple
-            rawPayload = payload,
-        )
-        val serialized = cose.serialize(ByteArraySerializer()).encodeToString(Base16Strict).uppercase()
-
-        serialized shouldContain "546869732069732074686520636F6E74656E742E" // "This is the content."
-        serialized shouldContain "43A10126"
-        cose.payload shouldBe payload
-    }
-
     "Serialization is correct with JSON" {
         val payload = "This is the content.".encodeToByteArray()
-        val cose = CoseSigned<ByteArray>(
-            protectedHeader = CoseHeader(algorithm = CoseAlgorithm.ES256),
+        val cose = CoseSigned.create(
+            protectedHeader = CoseHeader(algorithm = CoseAlgorithm.RS256),
             unprotectedHeader = CoseHeader(),
             payload = payload,
             signature = CryptoSignature.RSAorHMAC("bar".encodeToByteArray()),
-            rawPayload = payload,
+            payloadSerializer = ByteArraySerializer(),
         )
 
         Json.decodeFromString<CoseSigned<ByteArray>>(Json.encodeToString(cose)) shouldBe cose
@@ -64,39 +51,65 @@ class CoseSerializationTest : FreeSpec({
 
     "Serialization is correct with JSON for data class" {
         val payload = DataClass("This is the content.")
-        val cose = CoseSigned<DataClass>(
-            protectedHeader = CoseHeader(algorithm = CoseAlgorithm.ES256),
+        val cose = CoseSigned.create(
+            protectedHeader = CoseHeader(algorithm = CoseAlgorithm.RS256),
             unprotectedHeader = CoseHeader(),
             payload = payload,
             signature = CryptoSignature.RSAorHMAC("bar".encodeToByteArray()),
-            rawPayload = coseCompliantSerializer.encodeToByteArray(payload),
+            payloadSerializer = DataClass.serializer(),
         )
 
         Json.decodeFromString<CoseSigned<DataClass>>(Json.encodeToString(cose)) shouldBe cose
     }
 
-    "Serialization is correct for data class" {
-        val payload = DataClass("This is the content.")
-        val cose = CoseSigned(
-            protectedHeader = CoseHeader(algorithm = CoseAlgorithm.ES256),
+    "Serialization is correct for byte array" {
+        val payload = "This is the content.".encodeToByteArray()
+        val cose = CoseSigned.create(
+            protectedHeader = CoseHeader(algorithm = CoseAlgorithm.RS256),
             unprotectedHeader = CoseHeader(),
             payload = payload,
             signature = CryptoSignature.RSAorHMAC("bar".encodeToByteArray()), //RSAorHMAC because EC expects tuple
-            rawPayload = coseCompliantSerializer.encodeToByteArray(payload),
+            payloadSerializer = ByteArraySerializer(),
+        )
+        val serialized = cose.serialize(ByteArraySerializer()).encodeToString(Base16Strict).uppercase()
+
+        serialized shouldContain "546869732069732074686520636F6E74656E742E" // "This is the content."
+        serialized shouldContain "8445A101390100" // array of 5 bytes that is a map with -257 (the header for RS256)
+        cose.payload shouldBe payload
+    }
+
+    "Deserialization is correct for byte array" {
+        val input = "8445A101390100A054546869732069732074686520636F6E74656E742E43626172"
+
+        val cose = CoseSigned.deserialize(DataClass.serializer(), input.decodeToByteArray(Base16())).getOrThrow()
+        cose.payload shouldBe "This is the content.".encodeToByteArray()
+        cose.wireFormat.payload shouldBe "546869732069732074686520636F6E74656E742E".decodeToByteArray(Base16())
+        // important part is the D818 as tag(24)
+    }
+
+    "Serialization is correct for data class" {
+        val payload = DataClass("This is the content.")
+        val cose = CoseSigned.create(
+            protectedHeader = CoseHeader(algorithm = CoseAlgorithm.RS256),
+            unprotectedHeader = CoseHeader(),
+            payload = payload,
+            signature = CryptoSignature.RSAorHMAC("bar".encodeToByteArray()), //RSAorHMAC because EC expects tuple
+            payloadSerializer = DataClass.serializer(),
         )
         val serialized = cose.serialize(DataClass.serializer()).encodeToString(Base16Strict).uppercase()
 
         serialized shouldContain "546869732069732074686520636F6E74656E742E" // "This is the content."
-        serialized shouldContain "43A10126"
+        serialized shouldContain "8445A101390100" // array of 5 bytes that is a map with -257 (the header for RS256)
         cose.payload shouldBe payload
     }
 
     "Deserialization is correct for data class" {
-        val input = "8443A10126A0D818581EA167636F6E74656E7474546869732069732074686520636F6E74656E742E43626172"
+        val input = "8445A101390100A05822D818581EA167636F6E74656E7474546869732069732074686520636F6E74656E742E43626172"
 
         val cose = CoseSigned.deserialize(DataClass.serializer(), input.decodeToByteArray(Base16())).getOrThrow()
         cose.payload shouldBe DataClass("This is the content.")
-        cose.rawPayload shouldBe "D818581EA167636F6E74656E7474546869732069732074686520636F6E74656E742E".decodeToByteArray(Base16())
+        cose.wireFormat.payload shouldBe "D818581EA167636F6E74656E7474546869732069732074686520636F6E74656E742E"
+            .decodeToByteArray(Base16())
         // important part is the D818 as tag(24)
     }
 
@@ -123,7 +136,7 @@ class CoseSerializationTest : FreeSpec({
         deserialized.kid shouldBe header.kid
     }
 
-    "Deserialization is correct" {
+    "Deserialization is correct for unprotected header too" {
         val input = "d28443a10126a10442313154546869732069732074686520636f6e74656e" +
                 "742e58408eb33e4ca31d1c465ab05aac34cc6b23d58fef5c083106c4d25a" +
                 "91aef0b0117e2af9a291aa32e14ab834dc56ed2a223444547e01f11d3b09" +
@@ -140,19 +153,21 @@ class CoseSerializationTest : FreeSpec({
         val header = CoseHeader(algorithm = CoseAlgorithm.ES256)
         val inputManual = CoseSignatureInput(
             contextString = "Signature1",
-            protectedHeader = ByteStringWrapper(header),
+            protectedHeader = coseCompliantSerializer.encodeToByteArray(header),
             externalAad = byteArrayOf(),
             payload = payload
         ).serialize().encodeToString(Base16())
 
-        val inputLibrary = CoseSigned.prepareCoseSignatureInput(
+        val inputObject = CoseSigned.create(
             protectedHeader = header,
             payload = payload,
-            serializer = ByteArraySerializer(),
-        ).serialize().encodeToString(Base16())
+            signature = CryptoSignature.RSAorHMAC("bar".encodeToByteArray()),
+            payloadSerializer = ByteArraySerializer(),
+        ).prepareCoseSignatureInput(byteArrayOf())
+            .encodeToString(Base16())
 
         inputManual.shouldContain("Signature1".encodeToByteArray().encodeToString(Base16()))
-        inputLibrary shouldBe inputManual
+        inputObject shouldBe inputManual
     }
 
     "CoseSignatureInput is correct for custom types" {
@@ -160,20 +175,23 @@ class CoseSerializationTest : FreeSpec({
         val header = CoseHeader(algorithm = CoseAlgorithm.ES256)
         val inputManual = CoseSignatureInput(
             contextString = "Signature1",
-            protectedHeader = ByteStringWrapper(header),
+            protectedHeader = coseCompliantSerializer.encodeToByteArray(header),
             externalAad = byteArrayOf(),
-            payload = coseCompliantSerializer.encodeToByteArray(ByteStringWrapper(payload)),
+            payload = coseCompliantSerializer.encodeToByteArray(ByteStringWrapper(payload)).wrapInCborTag(24),
         ).serialize().encodeToString(Base16())
 
-        val inputLibrary = CoseSigned.prepareCoseSignatureInput(
+        val inputObject = CoseSigned.create(
             protectedHeader = header,
             payload = payload,
-            serializer = DataClass.serializer(),
-        ).serialize().encodeToString(Base16())
+            signature = CryptoSignature.RSAorHMAC("bar".encodeToByteArray()),
+            payloadSerializer = DataClass.serializer(),
+        ).prepareCoseSignatureInput(byteArrayOf())
+            .encodeToString(Base16())
 
         inputManual.shouldContain("Signature1".encodeToByteArray().encodeToString(Base16()))
-        inputLibrary shouldBe inputManual
+        inputObject shouldBe inputManual
     }
 
 
 })
+private fun ByteArray.wrapInCborTag(tag: Byte) = byteArrayOf(0xd8.toByte()) + byteArrayOf(tag) + this